+++ tests/unit/aggregation/test_constant.py
@@ -17,7 +17,7 @@ from ._inputs import non_strong_matrices, scaled_matrices, typical_matrices
 
 def _make_aggregator(matrix: Tensor) -> Constant:
     n_rows = matrix.shape[0]
-    weights = torch.tensor([1.0 / n_rows] * n_rows)
+    weights = torch.tensor([1.0 / n_rows] * n_rows, dtype=matrix.dtype)
     return Constant(weights)
